+++ packages/rune-components/src/combobox/previous/Previous.svelte
@@ -9,8 +9,11 @@
 	import { getRootContext } from '../context.js';
 	import { mergeProps } from '@svelte-thing/component-utils/svelte';
 
-	const { ...restProps }: HTMLButtonAttributes = $props();
-	const root = getRootContext<TRoot<unknown>>();
+	interface Props extends HTMLButtonAttributes {
+		root?: TRoot;
+	}
+
+	const { root = getRootContext(), ...restProps }: Props = $props();
 	const previous = createPrevious({
 		root,
 	} satisfies NullablyRequired<CreatePreviousConfig<unknown>>);