+++ packages/rune-components/src/menu/popup/Popup.svelte
@@ -1,14 +1,19 @@
 <script lang="ts">
 	import type { HTMLAttributes } from 'svelte/elements';
 	import type { NullablyRequired } from '../../types';
-	import { createPopup, type CreatePopupConfig } from './createPopup.svelte';
+	import type { TRoot } from '../root/createRoot.svelte.js';
+	import {
+		createPopup,
+		type CreatePopupConfig,
+	} from './createPopup.svelte.js';
+	import { getRootContext } from '../context.js';
 	import { mergeProps } from '@svelte-thing/component-utils/svelte';
-	import { getRootContext } from '../context';
 
-	type Props = HTMLAttributes<HTMLUListElement>;
+	interface Props extends HTMLAttributes<HTMLUListElement> {
+		root?: TRoot;
+	}
 
-	const { children, ...restProps }: Props = $props();
-	const root = getRootContext();
+	const { children, root = getRootContext(), ...restProps }: Props = $props();
 	const popup = createPopup({
 		root,
 	} satisfies NullablyRequired<CreatePopupConfig>);