+++ packages/rune-components/src/menu/trigger/Trigger.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import type { HTMLButtonAttributes } from 'svelte/elements';
 	import type { NullablyRequired } from '../../types.js';
+	import type { TRoot } from '../root/createRoot.svelte.js';
 	import {
 		createTrigger,
 		type CreateTriggerConfig,
@@ -8,8 +9,11 @@
 	import { getRootContext } from '../context.js';
 	import { mergeProps } from '@svelte-thing/component-utils/svelte';
 
-	const { children, ...restProps }: HTMLButtonAttributes = $props();
-	const root = getRootContext();
+	interface Props extends HTMLButtonAttributes {
+		root?: TRoot;
+	}
+
+	const { children, root = getRootContext(), ...restProps }: Props = $props();
 	const trigger = createTrigger({
 		root,
 	} satisfies NullablyRequired<CreateTriggerConfig>);