+++ sites/workspace/CHANGELOG.md
@@ -1,5 +1,12 @@
 # workspace-site
 
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies
+    - @svelte-thing/rune-components@0.0.25
+
 ## 0.0.39
 
 ### Patch Changes